refactor : reorganize user package structure
