+++ src/main/java/com/expense_tracker/user/repository/UserInfoRepository.java
@@ -1,9 +1,11 @@
-package com.expense_tracker.user;
+package com.expense_tracker.user.repository;
 
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+
+import com.expense_tracker.user.entity.UserInfo;
 
 @Repository
 public interface UserInfoRepository extends JpaRepository<UserInfo, Long> {
